python/cugraph/pyproject.toml
# Copyright (c) 2022, NVIDIA CORPORATION.

[build-system]

requires = [
    "wheel",
    "setuptools",
    "cython>=0.29,<0.30",
    "scikit-build>=0.13.1",
    "cmake>=3.23.1,!=3.25.0",
    "ninja",
    "rmm==23.6.*",
    "pylibraft==23.6.*",
    "pylibcugraph==23.6.*",
]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["cugraph/tests"]

[project]
name = "cugraph"
version = "23.06.00"
description = "cuGraph - RAPIDS GPU Graph Analytics"
readme = { file = "README.md", content-type = "text/markdown" }
authors = [
    { name = "NVIDIA Corporation" },
]
license = { text = "Apache 2.0" }
requires-python = ">=3.8"
dependencies = [
    "numba",
    "dask-cuda==23.6.*",
    "rmm==23.6.*",
    "cudf==23.6.*",
    "raft-dask==23.6.*",
    "dask-cudf==23.6.*",
    "pylibcugraph==23.6.*",
    "cupy-cuda11x",
]
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.10",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-xdist",
    "pytest-benchmark",
    "scipy",
    "numpy",
    "pandas",
    "networkx>=2.5.1",
    "scikit-learn>=0.23.1",
    "python-louvain",
    # cudf will use fsspec but is protocol independent. cugraph tests
    # specifically require http for the test files it asks cudf to read.
    "fsspec[http]>=0.6.0",
]

[project.urls]
Homepage = "https://github.com/rapidsai/cugraph"
Documentation = "https://docs.rapids.ai/api/cugraph/stable/"

[tool.setuptools]
license-files = ["LICENSE"]
